Improve addSignature logging to show required and current signatures for tx
Moved addSignature tests from BridgeSupportTestPowerMock to BridgeSupportTest

Removed unnecessary variables from some tests on BridgeSupportTestPowerMock
